In class GridLayoutImplYee : public GridLayoutImpl, private GridLayoutImplInternals I have implemented     virtual std::vector < std::tuple < uint32, Point> >     fieldNodeCoordinates1D( const Field & field, const Point & patchOrigin ) const override ;
this method returns a tuple with linear cell index and linear arrays containing point coordinates
for a given field.

I have added a new attribute in class GridLayoutImplInternals:
+ std::array< std::array<LayoutType,3>, 9 > hybridQtyCentering_ ;
it is filled in the ctor of GridLayoutImpl derived classes (like GridLayoutImplYee).

It provides the primal or dual nature of each component of an HybridQuantity (ex: Ex, Ey ...)
It is quite usefull and this is basically what we want the gridLayout to know about.