add 1 node to primal ghosts in 1st oder interpolation cases
Problem:
In first order interpolation cases, primal ghosts are not needed
for particle/mesh interpolation. Therefore the current version of the code
allocated NO ghost primal node. This prevent the Laplacian(J) to be calculated
This commit set +1 to primal ghosts in this case so that lap(J) can be calculated.